+++ app/cart/components/CartSummary.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import useCartStore from "@/app/product/cart/stores/useCartStore";
+import useCartStore from "@/app/cart/stores/useCartStore";
 import { formatPriceToKor } from "@/lib/utils";
 
 function CartSummary() {
